+++ changelog.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.4] - 2025-01-27
+## [1.0.6] - 2025-06-23
+
+### Fixed
+- Fixed using wrong models to represent the Battle log.
+
+## [1.0.5] - 2025-06-23
+
+### Changed
+- Models are now exported too
+
+## [1.0.4] - 2025-06-21
 
 ### Fixed
 - Fixed all `__init__.py` files to properly expose library classes and modules.